Move HiveMetaStoreClientFactory from hive-exec to standalone-metastore-common
